feat: working Sonar Issues and requested changes
